flask-template/requirements.txt
flask
flask-restful
flask-session
langchain
langchain_community
langchain_huggingface
PyPDF2
python-dotenv
faiss-cpu
python-docx
langchain-google-genai
waitress
google-cloud-storage